+++ packages/ui/src/index.ts
@@ -9,3 +9,11 @@ export * from './Dialog';
 export * from './Table';
 export * from './Tabs';
 export * from './Tooltip';
+
+export * from './Autocomplete';
+export * from './Fab';
+export * from './IconButton';
+export * from './Radio';
+export * from './Rating';
+export * from './Slider';
+export * from './ToggleButton';